+++ simple-stack/src/test/java/com/zhuinden/simplestack/FlowTest.java
@@ -25,99 +25,23 @@ import org.junit.Test;
 
 import java.util.Arrays;
 import java.util.Iterator;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.MockitoAnnotations.initMocks;
 
 public class FlowTest {
-    static class Uno
-            implements Parcelable {
+    static class Uno {
         public Uno() {
         }
-
-        protected Uno(Parcel in) {
-        }
-
-        public static final Creator<Uno> CREATOR = new Creator<Uno>() {
-            @Override
-            public Uno createFromParcel(Parcel in) {
-                return new Uno(in);
-            }
-
-            @Override
-            public Uno[] newArray(int size) {
-                return new Uno[size];
-            }
-        };
-
-        @Override
-        public int describeContents() {
-            return 0;
-        }
-
-        @Override
-        public void writeToParcel(Parcel dest, int flags) {
-        }
     }
 
-    static class Dos
-            implements Parcelable {
+    static class Dos {
         public Dos() {
         }
-
-        protected Dos(Parcel in) {
-        }
-
-        public static final Creator<Dos> CREATOR = new Creator<Dos>() {
-            @Override
-            public Dos createFromParcel(Parcel in) {
-                return new Dos(in);
-            }
-
-            @Override
-            public Dos[] newArray(int size) {
-                return new Dos[size];
-            }
-        };
-
-        @Override
-        public int describeContents() {
-            return 0;
-        }
-
-        @Override
-        public void writeToParcel(Parcel dest, int flags) {
-        }
     }
 
-    static class Tres
-            implements Parcelable {
+    static class Tres {
         public Tres() {
-        }
-
-        protected Tres(Parcel in) {
-        }
-
-        public static final Creator<Tres> CREATOR = new Creator<Tres>() {
-            @Override
-            public Tres createFromParcel(Parcel in) {
-                return new Tres(in);
-            }
-
-            @Override
-            public Tres[] newArray(int size) {
-                return new Tres[size];
-            }
-        };
-
-        @Override
-        public int describeContents() {
-            return 0;
-        }
-
-        @Override
-        public void writeToParcel(Parcel dest, int flags) {
         }
     }
 
@@ -126,7 +50,7 @@ public class FlowTest {
     final TestKey charlie = new TestKey("Charlie");
     final TestKey delta = new TestKey("Delta");
 
-    List<?> lastStack;
+    History<?> lastStack;
     int lastDirection;
 
     class FlowDispatcher
@@ -146,24 +70,24 @@ public class FlowTest {
 
     @Test
     public void oneTwoThree() {
-        List<?> history = History.single(new Uno());
+        History<?> history = History.single(new Uno());
         Backstack flow = new Backstack(history);
         flow.setStateChanger(new FlowDispatcher());
 
         flow.goTo(new Dos());
-        assertThat(lastStack.get(lastStack.size() - 1)).isInstanceOf(Dos.class);
+        assertThat(lastStack.top()).isInstanceOf(Dos.class);
         assertThat(lastDirection).isSameAs(StateChange.FORWARD);
 
         flow.goTo(new Tres());
-        assertThat(lastStack.get(lastStack.size() - 1)).isInstanceOf(Tres.class);
+        assertThat(lastStack.top()).isInstanceOf(Tres.class);
         assertThat(lastDirection).isSameAs(StateChange.FORWARD);
 
         assertThat(flow.goBack()).isTrue();
-        assertThat(lastStack.get(lastStack.size() - 1)).isInstanceOf(Dos.class);
+        assertThat(lastStack.top()).isInstanceOf(Dos.class);
         assertThat(lastDirection).isSameAs(StateChange.BACKWARD);
 
         assertThat(flow.goBack()).isTrue();
-        assertThat(lastStack.get(lastStack.size() - 1)).isInstanceOf(Uno.class);
+        assertThat(lastStack.top()).isInstanceOf(Uno.class);
         assertThat(lastDirection).isSameAs(StateChange.BACKWARD);
 
         assertThat(flow.goBack()).isFalse();
@@ -171,7 +95,7 @@ public class FlowTest {
 
     @Test
     public void historyChangesAfterListenerCall() {
-        final List<?> firstHistory = History.single(new Uno());
+        final History<?> firstHistory = History.single(new Uno());
 
         class Ourrobouros
                 implements StateChanger {
@@ -198,56 +122,56 @@ public class FlowTest {
 
     @Test
     public void historyPushAllIsPushy() {
-        List<?> history = History.from(Arrays.asList(able, baker, charlie));
+        History<?> history = History.from(Arrays.asList(able, baker, charlie));
         assertThat(history.size()).isEqualTo(3);
 
         Backstack flow = new Backstack(history);
         flow.setStateChanger(new FlowDispatcher());
 
         assertThat(flow.goBack()).isTrue();
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(baker);
+        assertThat(lastStack.top()).isEqualTo(baker);
 
         assertThat(flow.goBack()).isTrue();
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(able);
+        assertThat(lastStack.top()).isEqualTo(able);
 
         assertThat(flow.goBack()).isFalse();
     }
 
     @Test
     public void setHistoryWorks() {
-        List<?> history = History.from(Arrays.asList(able, baker));
+        History<?> history = History.from(Arrays.asList(able, baker));
         Backstack flow = new Backstack(history);
         FlowDispatcher handleStateChangeer = new FlowDispatcher();
         flow.setStateChanger(handleStateChangeer);
 
-        List<?> newHistory = History.from(Arrays.asList(charlie, delta));
+        History<?> newHistory = History.from(Arrays.asList(charlie, delta));
         flow.setHistory(newHistory, StateChange.FORWARD);
         assertThat(lastDirection).isSameAs(StateChange.FORWARD);
-        assertThat(lastStack.get(lastStack.size() - 1)).isSameAs(delta);
+        assertThat(lastStack.top()).isSameAs(delta);
         assertThat(flow.goBack()).isTrue();
-        assertThat(lastStack.get(lastStack.size() - 1)).isSameAs(charlie);
+        assertThat(lastStack.top()).isSameAs(charlie);
         assertThat(flow.goBack()).isFalse();
     }
 
     @Test
     public void setObjectGoesBack() {
-        List<?> history = History.from(Arrays.asList(able, baker, charlie, delta));
+        History<?> history = History.from(Arrays.asList(able, baker, charlie, delta));
         Backstack flow = new Backstack(history);
         flow.setStateChanger(new FlowDispatcher());
 
         assertThat(history.size()).isEqualTo(4);
 
         flow.goTo(charlie);
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(charlie);
+        assertThat(lastStack.top()).isEqualTo(charlie);
         assertThat(lastStack.size()).isEqualTo(3);
         assertThat(lastDirection).isEqualTo(StateChange.BACKWARD);
 
         assertThat(flow.goBack()).isTrue();
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(baker);
+        assertThat(lastStack.top()).isEqualTo(baker);
         assertThat(lastDirection).isEqualTo(StateChange.BACKWARD);
 
         assertThat(flow.goBack()).isTrue();
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(able);
+        assertThat(lastStack.top()).isEqualTo(able);
         assertThat(lastDirection).isEqualTo(StateChange.BACKWARD);
 
         assertThat(flow.goBack()).isFalse();
@@ -255,67 +179,67 @@ public class FlowTest {
 
     @Test
     public void setObjectToMissingObjectPushes() {
-        List<?> history = History.from(Arrays.asList(able, baker));
+        History<?> history = History.from(Arrays.asList(able, baker));
         Backstack flow = new Backstack(history);
         flow.setStateChanger(new FlowDispatcher());
         assertThat(history.size()).isEqualTo(2);
 
         flow.goTo(charlie);
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(charlie);
+        assertThat(lastStack.top()).isEqualTo(charlie);
         assertThat(lastStack.size()).isEqualTo(3);
         assertThat(lastDirection).isEqualTo(StateChange.FORWARD);
 
         assertThat(flow.goBack()).isTrue();
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(baker);
+        assertThat(lastStack.top()).isEqualTo(baker);
         assertThat(lastDirection).isEqualTo(StateChange.BACKWARD);
 
         assertThat(flow.goBack()).isTrue();
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(able);
+        assertThat(lastStack.top()).isEqualTo(able);
         assertThat(lastDirection).isEqualTo(StateChange.BACKWARD);
         assertThat(flow.goBack()).isFalse();
     }
 
     @Test
     public void setObjectKeepsOriginal() {
-        List<?> history = History.from(Arrays.asList(able, baker));
+        History<?> history = History.from(Arrays.asList(able, baker));
         Backstack flow = new Backstack(history);
         flow.setStateChanger(new FlowDispatcher());
         assertThat(history.size()).isEqualTo(2);
 
         flow.goTo(new TestKey("Able"));
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(new TestKey("Able"));
-        assertThat(lastStack.get(lastStack.size() - 1) == able).isTrue();
-        assertThat(lastStack.get(lastStack.size() - 1)).isSameAs(able);
+        assertThat(lastStack.top()).isEqualTo(new TestKey("Able"));
+        assertThat(lastStack.top() == able).isTrue();
+        assertThat(lastStack.top()).isSameAs(able);
         assertThat(lastStack.size()).isEqualTo(1);
         assertThat(lastDirection).isEqualTo(StateChange.BACKWARD);
     }
 
     @Test
     public void replaceHistoryResultsInLengthOneHistory() {
-        List<?> history = History.from(Arrays.asList(able, baker, charlie));
+        History<?> history = History.from(Arrays.asList(able, baker, charlie));
         Backstack flow = new Backstack(history);
         flow.setStateChanger(new FlowDispatcher());
         assertThat(history.size()).isEqualTo(3);
 
         flow.setHistory(History.single(delta), StateChange.REPLACE);
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(new TestKey("Delta"));
-        assertThat(lastStack.get(lastStack.size() - 1) == delta).isTrue();
-        assertThat(lastStack.get(lastStack.size() - 1)).isSameAs(delta);
+        assertThat(lastStack.top()).isEqualTo(new TestKey("Delta"));
+        assertThat(lastStack.top() == delta).isTrue();
+        assertThat(lastStack.top()).isSameAs(delta);
         assertThat(lastStack.size()).isEqualTo(1);
         assertThat(lastDirection).isEqualTo(StateChange.REPLACE);
     }
 
     @Test
     public void replaceTopDoesNotAlterHistoryLength() {
-        List<?> history = History.from(Arrays.asList(able, baker, charlie));
+        History<?> history = History.from(Arrays.asList(able, baker, charlie));
         Backstack flow = new Backstack(history);
         flow.setStateChanger(new FlowDispatcher());
         assertThat(history.size()).isEqualTo(3);
 
         flow.setHistory(History.builderFrom(flow).removeLast().add(delta).build(), StateChange.REPLACE);
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(new TestKey("Delta"));
-        assertThat(lastStack.get(lastStack.size() - 1) == delta).isTrue();
-        assertThat(lastStack.get(lastStack.size() - 1)).isSameAs(delta);
+        assertThat(lastStack.top()).isEqualTo(new TestKey("Delta"));
+        assertThat(lastStack.top() == delta).isTrue();
+        assertThat(lastStack.top()).isSameAs(delta);
         assertThat(lastStack.size()).isEqualTo(3);
         assertThat(lastDirection).isEqualTo(StateChange.REPLACE);
     }
@@ -327,26 +251,26 @@ public class FlowTest {
         TestKey baker = new TestKey("Baker");
         TestKey charlie = new TestKey("Charlie");
         TestKey delta = new TestKey("Delta");
-        List<?> history = History.from(Arrays.asList(able, baker, charlie, delta));
+        History<?> history = History.from(Arrays.asList(able, baker, charlie, delta));
         Backstack flow = new Backstack(history);
         flow.setStateChanger(new FlowDispatcher());
         assertThat(history.size()).isEqualTo(4);
 
         TestKey echo = new TestKey("Echo");
         TestKey foxtrot = new TestKey("Foxtrot");
-        List<?> newHistory = History.from(Arrays.asList(able, baker, echo, foxtrot));
+        History<?> newHistory = History.from(Arrays.asList(able, baker, echo, foxtrot));
         flow.setHistory(newHistory, StateChange.REPLACE);
         assertThat(lastStack.size()).isEqualTo(4);
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(foxtrot);
+        assertThat(lastStack.top()).isEqualTo(foxtrot);
         flow.goBack();
         assertThat(lastStack.size()).isEqualTo(3);
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(echo);
+        assertThat(lastStack.top()).isEqualTo(echo);
         flow.goBack();
         assertThat(lastStack.size()).isEqualTo(2);
-        assertThat(lastStack.get(lastStack.size() - 1)).isSameAs(baker);
+        assertThat(lastStack.top()).isSameAs(baker);
         flow.goBack();
         assertThat(lastStack.size()).isEqualTo(1);
-        assertThat(lastStack.get(lastStack.size() - 1)).isSameAs(able);
+        assertThat(lastStack.top()).isSameAs(able);
     }
 
     static class Picky
@@ -404,7 +328,7 @@ public class FlowTest {
 
     @Test
     public void setCallsEquals() {
-        List<?> history = History.newBuilder()
+        History<?> history = History.newBuilder()
                 .addAll(Arrays.asList(new Picky("Able"), new Picky("Baker"), new Picky("Charlie"), new Picky("Delta")))
                 .build();
         Backstack flow = new Backstack(history);
@@ -413,16 +337,16 @@ public class FlowTest {
         assertThat(history.size()).isEqualTo(4);
 
         flow.goTo(new Picky("Charlie"));
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(new Picky("Charlie"));
+        assertThat(lastStack.top()).isEqualTo(new Picky("Charlie"));
         assertThat(lastStack.size()).isEqualTo(3);
         assertThat(lastDirection).isEqualTo(StateChange.BACKWARD);
 
         assertThat(flow.goBack()).isTrue();
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(new Picky("Baker"));
+        assertThat(lastStack.top()).isEqualTo(new Picky("Baker"));
         assertThat(lastDirection).isEqualTo(StateChange.BACKWARD);
 
         assertThat(flow.goBack()).isTrue();
-        assertThat(lastStack.get(lastStack.size() - 1)).isEqualTo(new Picky("Able"));
+        assertThat(lastStack.top()).isEqualTo(new Picky("Able"));
         assertThat(lastDirection).isEqualTo(StateChange.BACKWARD);
 
         assertThat(flow.goBack()).isFalse();